Rename common.ts -> http-transport.ts
